requirements.txt
# Core CrewAI framework
crewai>=0.70.1
crewai[tools]>=0.70.1

# LLM and AI dependencies
openai>=1.51.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Web scraping and data processing
beautifulsoup4>=4.12.3
requests>=2.32.3
selenium>=4.26.1
scrapy>=2.11.2
lxml>=5.3.0

# Natural language processing
nltk>=3.9.1
spacy>=3.8.2
textblob>=0.18.0
transformers>=4.46.0

# Data analysis and visualization
pandas>=2.2.3
numpy>=2.1.3
matplotlib>=3.9.2
seaborn>=0.13.2
plotly>=5.24.1

# Web UI frameworks
streamlit>=1.39.0
gradio>=5.6.0

# Testing and development
pytest>=8.3.3
pytest-cov>=5.0.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
black>=24.10.0
flake8>=7.1.1
mypy>=1.13.0

# Logging and monitoring
loguru>=0.7.2
python-dotenv>=1.0.1

# Security and validation
pydantic>=2.9.2
validators>=0.34.0

# Rate limiting and caching
ratelimit>=2.2.1
redis>=5.1.1
cachetools>=5.5.0

# Database (optional for persistence)
sqlalchemy>=2.0.36
sqlite3

# Email notifications (optional)
smtplib

# Additional utilities
tqdm>=4.67.0
rich>=13.9.4
typer>=0.12.5
pyyaml>=6.0.2
